pyproject.toml
[project]
name = "gemini-mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.15.0",
    "mcp>=1.9.0",
    "python-dotenv>=1.1.0",
]
